refactor(ui): split AdminPage.jsx (1726 lines) into per-section files
AdminPage.jsx was a 1726-line god-component, well past the project's 1000-line
must-split rule. Extract each independent admin section into its own file under
components/admin/: PowerQueryExportSection, CuratedDataSection,
HistoryRetentionSection, DangerZoneSection, LLMSettingsSection, and the risk
cluster (RiskProfileSection/ClassifiersSection/…) as RiskScoringSection — plus
shared primitives in adminUi.jsx and the local date formatter in adminFormat.js.
AdminPage.jsx is now a ~166-line tab router.

Pure relocation, no behavioural change: AdminPage.mount.test.jsx (13 cases,
navigates every tab) and the updated source-scan AdminPage.test.js both pass;
eslint clean; import-conventions test green. Re-baselined the cyclomatic ratchet
for the one grandfathered-over-threshold unit that changed files
(LLMSettingsSection, CC 36) — the complexity is unchanged, only its location
(AdminPage.jsx -> admin/LLMSettingsSection.jsx).
